+++ CHANGELOG.md
@@ -21,6 +21,7 @@ first, also the PHP version required is 7.1, see [#426].
 - Add support to fill Mentioned on data (@glensc, #447, #449)
 - Listen for Issue/Note GitLab events (@glensc, #337, #448)
 - xmlrpc: verify token first to avoid incrementing failed logins (@glensc, #463)
+- Do issue duplication and close in same step (@glensc, #127)
 
 [3.6.0]: https://github.com/eventum/eventum/compare/v3.5.6...master
 [#426]: https://github.com/eventum/eventum/pull/426